BoardPage is now listening to the subscription hook. deletColumnById is being given the boardId so the subscription will work properly.

+++ src/graphql/column/hooks/useColumnDeleted.js
@@ -1,6 +1,6 @@
 import { useSubscription } from '@apollo/client'
 import { COLUMN_DELETED } from '../columnQueries'
-import { deleteColumnFromCache } from '../cacheService/cacheUpdates'
+import { deleteColumnFromCache } from '../../../cacheService/cacheUpdates'
 
 const useColumnDeleted = (id) => {
     useSubscription(COLUMN_DELETED,
@@ -8,7 +8,10 @@ const useColumnDeleted = (id) => {
             variables: { boardId: id },
             onSubscriptionData: ({ subscriptionData: { data } }) => {
                 const { columnId, boardId } = data.columnDeleted.removeInfo
-                deleteColumnFromCache(columnId, boardId)
+                if (data.columnDeleted.removeType === 'DELETED') {
+                    deleteColumnFromCache(columnId, boardId)
+                }
+
             }
         })
 }
